add: move shop types and add bundle util
