+++ src/utils/getBundles.ts
@@ -14,6 +14,7 @@ import type {
   BundledMysticCode
 } from "src/items/types";
 import type { BundledEvent } from "src/events/types";
+import type { BundledShops } from "src/shops/types";
 import { Log } from "./log";
 import { safeProxyDataMap } from "./safeProxyDataMap";
 
@@ -123,6 +124,10 @@ export const getCustomItems = createBundle<PartialDataMap<BundledItem>>(
 export const getCustomItemMap = withProxy(
   getCustomItems,
   "Could not find Custom Item %KEY% in bundled data"
+);
+
+export const getBundledShops = createBundle<BundledShops>(
+  path.join(process.cwd(), "assets/static/shops.json")
 );
 
 export const getBundledMysticCodes = createBundle<
